Moving ObjectCycle classes
