+++ Bukkit/src/main/java/me/oczi/bukkit/PluginCore.java
@@ -2,7 +2,7 @@ package me.oczi.bukkit;
 
 import me.oczi.bukkit.internal.CooldownManager;
 import me.oczi.bukkit.internal.MemoryManager;
-import me.oczi.bukkit.internal.ObjectCycleManager;
+import me.oczi.bukkit.internal.objectcycle.ObjectCycleManager;
 import me.oczi.bukkit.internal.commandflow.CommandFlow;
 import me.oczi.bukkit.internal.database.DatabaseManager;
 import me.oczi.bukkit.internal.database.DbTasks;
