+++ Bukkit/src/main/java/me/oczi/bukkit/internal/objectcycle/ObjectCycleManager.java
@@ -1,4 +1,4 @@
-package me.oczi.bukkit.internal;
+package me.oczi.bukkit.internal.objectcycle;
 
 import me.oczi.bukkit.internal.objectcycle.partner.PartnerObjectLoader;
 import me.oczi.bukkit.internal.objectcycle.player.PlayerObjectLoader;
